chore: extracted types from service
