+++ frontend/src/types/OpeningTypes.ts
@@ -1,0 +1,37 @@
+export interface statusCategory {
+  code: string;
+  description: string;
+}
+
+export interface RecentOpeningApi {
+  openingId: number;
+  forestFileId: string;
+  cuttingPermit: string | null;
+  timberMark: string | null;
+  cutBlock: string | null;
+  grossAreaHa: number | null;
+  status: statusCategory | null;
+  category: statusCategory | null;
+  disturbanceStart: string | null;
+  entryTimestamp: string | null;
+  updateTimestamp: string | null;
+}
+
+export interface IOpeningPerYear {
+  orgUnitCode: string | null;
+  statusCode: string | null;
+  entryDateStart: string | null;
+  entryDateEnd: string | null;
+}
+
+export interface IFreeGrowingProps {
+  orgUnitCode: string;
+  clientNumber: string;
+  entryDateStart: string | null;
+  entryDateEnd: string | null;
+}
+
+export interface IFreeGrowingChartData {
+  group: string;
+  value: number;
+}
